requirements.txt
langchain==0.0.287
openai==0.27.6
tiktoken==0.4.0
streamlit==1.22.0
faiss-cpu==1.7.4
streamlit-chat==0.0.2.2
pygpt4all==1.1.0
chromadb==0.4.14
llama-cpp-python==0.1.50
pdfminer.six==20221105
python-dotenv==1.0.0
unstructured==0.10.18
extract-msg==0.41.1
tabulate==0.9.0
pandoc==2.3
pinecone-client
pypandoc==1.11
pymupdf==1.22.5
pypdf==3.16.2
transformers==4.32.0
torch==2.0.1
llama-index==0.8.26
pinecone-text==0.5.4
nebula3-python==3.4.0
ipython-ngql==0.7.5
ghostscript==0.7
opencv-python==4.8.1.78
sqlalchemy==2.0.21
matplotlib==3.7.3
seaborn==0.13.0
pandas==1.5.3
matplotlib==3.7.3
seaborn==0.13.0
pysqlite3-binary==0.5.2
.